+++ CHANGES.rst
@@ -5,6 +5,7 @@
 ------------------
 
 - Backward compatibility for the schema < 300
+- Added function to check if schematisation DB file is actually a 3Di schematisation data source.
 
 0.1.6.dev0 (2025-02-13)
 -----------------------